개인 프로젝트

+++ college/app-server/src/main/java/college/pj/handler/MemberDeleteServlet.java
@@ -1,11 +1,13 @@
 package college.pj.handler;
 
 import java.io.IOException;
+import java.io.PrintWriter;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import college.pj.vo.Member;
 
 
 @WebServlet("/member/delete")
@@ -16,6 +18,17 @@ public class MemberDeleteServlet extends HttpServlet {
   @Override
   protected void doGet(HttpServletRequest request, HttpServletResponse response)
       throws ServletException, IOException {
+
+    Member loginUser = (Member) request.getSession().getAttribute("loginUser");
+
+    response.setContentType("text/html;charset=UTF-8");
+    PrintWriter out = response.getWriter();
+
+    if (loginUser.getNo() != Integer.parseInt(request.getParameter("no"))  && loginUser.getLevel() != 2) {
+      out.println("<p>권한이 없습니다.</p>");
+      out.println("<meta http-equiv='refresh' content='1;url=/member/list'>");
+      return;
+    }
 
     try {
       if (InitServlet.memberDao.delete(Integer.parseInt(request.getParameter("no"))) == 0) {
